tests: use appropriate return type for gethugepagesize()
While running `make func` on ia64, I found that
truncate_sigbus_versus_oom was adding a negative value to a pointer and
causing a SEGV. This was because hpage_size was of type int, even though
the return type of gethugepagesize() is long. The multiplication was
truncated and led to negative values. Fix this throughout the tests,
just in case, by making hpage_size always a long.

+++ elflink.c
@@ -539,7 +539,7 @@ static void parse_elf(Elf_Ehdr *ehdr)
  */
 static int prepare_segment(struct seg_info *seg)
 {
-	int hpage_size = gethugepagesize();
+	long hpage_size = gethugepagesize();
 	void *p;
 	unsigned long gap;
 	unsigned long size;
