+++ tests/gethugepagesize.c
@@ -26,14 +26,14 @@
 
 int main(int argc, char *argv[])
 {
-	int hpage_size;
+	long hpage_size;
 
 	test_init(argc, argv);
 
 	hpage_size = gethugepagesize();
 
 	if (hpage_size > 0) {
-		verbose_printf("Huge page size is %d bytes\n", hpage_size);
+		verbose_printf("Huge page size is %ld bytes\n", hpage_size);
 		PASS();
 	}
