+++ tests/private.c
@@ -32,7 +32,7 @@
 
 int main(int argc, char *argv[])
 {
-	int hpage_size;
+	long hpage_size;
 	int fd;
 	void *p, *q;
 	unsigned int *pl, *ql;
